+++ web/src/routes/(hp)/en/case/+layout.svelte
@@ -2,7 +2,7 @@
     import Footer from '$components/footer.svelte';
     import Nav from '$components/nav.svelte';
     import { setContext } from 'svelte';
-    import ContactDanielEn from '../contactDanielEn.svelte';
+    import ContactDanielEn from '../../../../components/contactDanielEn.svelte';
 
     setContext('lang', 'en');
 </script>
